removed async waits for simple/station queries because it was a problem for tests and not a necessary impl, refactored station into its own package, migrated shared classes from simplestation to the webservice package
