--- src/main/java/gov/usgs/cida/wqp/webservice/SimpleStation/IWqxOutbound.java
@@ -1,6 +1,0 @@
-package gov.usgs.cida.wqp.webservice.SimpleStation;
-
-public interface IWqxOutbound {
-
-	
-}
